testing

+++ application1/handler/data/reader/h5.py
@@ -13,7 +13,7 @@ LOG = ConfigurationManager.get_logger(__name__)
 
 class H5Reader(BaseReader):
 
-    RECORD_STRUCTURE = [('file', str, 100), ('gps_start', int), ('gps_end', int)]
+    RECORD_STRUCTURE = [('file', str, 100), ('gps_start', float), ('gps_end', float)]
     H5_DIR = 'ds_data/data/'
     H5 = '.h5'
 
@@ -34,7 +34,7 @@ class H5Reader(BaseReader):
         records = []
         for file in h5_files:
             _, gps_start, gps_end = split_file_name(file)
-            records.append((file, float(gps_start), float(gps_end)))
+            records.append((file, gps_start, gps_end))
         records = np.array(records, dtype=self.RECORD_STRUCTURE)
         records = records.view(dtype=(np.record, records.dtype), type=np.recarray)
         return records[(records.gps_end > self.gps_start) & (records.gps_start < self.gps_end)]
